feat: soporte para ordenes fraccionales (fractional shares)
Cambios clave:
- quantity pasa de int a float en Trade, OrderPreviewRequest, IBKRClient
- Elimina int() en calculo de units -> ahora usa round(units, 4)
- DB schema: trades.quantity INTEGER -> REAL
- Validator: position_size_units ya no se trunca
- Permite comprar exactamente el monto maximo permitido sin desperdiciar capital

+++ app/db/database.py
@@ -124,7 +124,7 @@ def init_db():
             CREATE TABLE IF NOT EXISTS trades (
                 id INTEGER PRIMARY KEY AUTOINCREMENT,
                 symbol TEXT NOT NULL, action TEXT NOT NULL,
-                quantity INTEGER NOT NULL, entry_price REAL NOT NULL,
+                quantity REAL NOT NULL, entry_price REAL NOT NULL,
                 stop_loss_price REAL NOT NULL, take_profit_price REAL NOT NULL,
                 stop_loss_pct REAL NOT NULL, take_profit_pct REAL NOT NULL,
                 signal_strength TEXT NOT NULL, llm_justification TEXT,
